requirements.txt
keyboard==0.13.5
sounddevice==0.4.6
numpy==1.24.3
faster-whisper==1.0.3
pyperclip==1.8.2
scipy==1.11.4
pystray==0.19.5
Pillow==10.1.0
